Add command line parsing and silence a good number of compiler warnings.
A lot of warnings were mostly for things like unused functions. Most of
the offenders have either been removed (because they're easy enough to
re-write) or moved to only be built during testing. The rest I have kept
because I believe I will use them in the future when compiling for the
VM.

Command line parsing is also here, where it will parse a file from the
CLI and print out its AST.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>