Add bytecode and compiler.
Bytecode and compiler are relatively small pieces of this actual
language, so they can all be smushed into this commit. Lots of changes
across the board, especially in the VM module. Compiler *should* be
emitting correct bytecode, I haven't had a chance to dump the bytecode
and give it a look-over.

Beyond that there's a new `matches` package included, which is just a
couple of useful macros.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>