Update packages and add builtins.
The test program is able to run now, even though it's only able to get
through about 50% of it before it breaks due to some builtins not being
implemented. However, there ARE some builtins, which is a good start.

lazy_static was running into issues with the new Rust syntax for 1.18
with visibility restrictions. I'm using my fork with the fix until it
gets merged into the official branch.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>