Add preprocessing and basic.sbl
Preprocessing had a lot of things to change in various files, including
a method for reading an input path and turning it into an AST.

The old basic.sbl file has been added to the repo.

With builtins, I've started experimenting with the idea of "consumers"
and "observers"; methods which pop and peek the top item(s) of the
stack, respectively.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>