Add start for FFI
FFI stuff is kind of a challenge because we have to figure out the best
way to interface with C, especially with both mutable and const stacks
in order to act like arrays.

This commit adds the beginnings of a FFI to SBL; this commit just adds
the backbone and ability to call foreign functions with (most) SBL
types. These aren't even being parsed yet, so that will have to come
next.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>