Add string pool to foreign function calls
Strings are a little weird because they need to go through three
transformations, and have some odd contingencies (e.g. we need a
reference to a pointer in the end... it's very odd.) The best way to
solve this, in my opinion, is with a memory pool that just holds
ownership to CStrings (which add a null byte at the end of them) so we
can get those pointer references. A little complicated, but it works.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>