Foreign (#9)
* Add start for FFI

FFI stuff is kind of a challenge because we have to figure out the best
way to interface with C, especially with both mutable and const stacks
in order to act like arrays.

This commit adds the beginnings of a FFI to SBL; this commit just adds
the backbone and ability to call foreign functions with (most) SBL
types. These aren't even being parsed yet, so that will have to come
next.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Fix small compilation error

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Add foreign parsing and tests

Foreign functions are now being added to the AST. They have not yet been
added to the VM, but that is the next step.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Update functions to follow unified structure

This cleans up the code a little bit and makes everything less hacky all
around.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Fix up stuff to be able to execute foreign functions.

So far, I haven't gotten a foreign function to execute correctly....
yet. But we're getting there.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Minor refactoring in builtins.rs (#8)

* Minor refactoring.

* Completing builtins refactoring.

* Add string pool to foreign function calls

Strings are a little weird because they need to go through three
transformations, and have some odd contingencies (e.g. we need a
reference to a pointer in the end... it's very odd.) The best way to
solve this, in my opinion, is with a memory pool that just holds
ownership to CStrings (which add a null byte at the end of them) so we
can get those pointer references. A little complicated, but it works.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Add small IO library that tests use of foreign functions

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Add travis-ci

* Add texinfo dependency to travis build for libffi

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Readme (#12)

* Add README and CONTRIBUTING. Closes #6

* Add GRAMMAR.md and updated it to match current syntax

* Update links to FFI issues in README

* Update GRAMMAR.md to be more accurate for the current grammar

* Turn off all email notifications for CI

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Ran rustfmt on all source

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Squash a bunch of warnings, making the compile output more meaningful

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Remove commented impl that was causing tests to fail

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>