Optimizer (#19)
* Re-organize compiler

* Add optimization skeleton with function inlining.

Functions that do not call other functions are now being inlined. There
is still some work to do but this is a big step as far as optimizations
go.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Add optimization flag and ran rustfmt on the project again

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Remove unused function

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>

* Remove snip_inlines from the function inliner

This is future-proofing for compiled files. If functions are getting
snipped because they're being inlined, and then they get called by an
external file and they aren't there, that is a problem. So inlined
functions are kept in.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>