Decouple block compilation behavior from the compiler
Blocks of code up to this point have been compiled via a private method
in the CompileBytes impl. However, we need to separate this behavior in
order to implement bake blocks without repeating code. The CompileBlock
struct has been created for this, and is used to compile individual
blocks of code (given a BoringTable ref, a Block ref, and an offset).

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>