Add compile-time bake blocks.
This is a big(ger) commit, but it gives us a good amount of behavior
I've been wanting in the compiler. Bake-blocks are allowed anywhere
executable code goes, and during compile time these blocks are executed,
resulting in push instructions for each value that was left on the stack
on exit.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>