Add IR, moving bytecode to be on the VM end.
Intermediate representation is specific to the compiler. This helps
solve the problem of bake blocks needing to be a VM construct, even
though they're a language construct. Bake blocks are no longer
compiling, and the bake.rs file is no longer in place because of this.
However, this was a necessary refactor because of how complicated and
split the VM and compiler are becoming; they are no longer 1:1
especially during compile-time.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>