Add labels instead of hard-addressing
The most recent optimization revealed a serious bug that I had
overlooked regarding jump addresses. Basically, because optimizations
fool around with the instructions, the jump instructions' jump labels
become out-of-date and the program erroneously jumps to the wrong
locations. I'm in the middle of adding labels, and while it compiles,
the programs are still incorrect.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>