Add STORE instruction, and optimization reorg
STORE instruction has been added, which converts a PUSH followed
immediately by a POP into a single instruction which pops the top value
off of the stack and stores it in the given variable. This is an
optimization, run right before PUSH_COMPRESS.

Optimizations have been reorganized a little bit, so they all live in
the optimize module while still being part of the BCFun impl. They are
also only public in the optimize module.

Everything else falls in the scope of minor changes (e.g. dump output
formatting).

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>