Fix elbr statement bug mentioned in the last commit
If the final elbr statement in a br/elbr chain was not followed by an el
statement, its final "cleanup" bit wasn't getting filled in, leaving an
unexpected item behind on the stack.

This has been fixed, and I've added a couple of debug_assert!s that will
catch whenever a NOP is left in (but only during debug builds, checking
every time can be expensive).

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>